+++ src/h5fort.F90
@@ -9,6 +9,7 @@ module h5fort
     H5FORTRAN_READ_ONLY_PARALLEL => H5FORTRAN_READ_ONLY
 #endif
 #ifdef H5FORTRAN_SERIAL
+  use h5fort_serial_attribute, only: h5fort_get_attribute_info
   use h5fort_serial, H5FORTRAN_FORCE_WRITE_SERIAL => H5FORTRAN_FORCE_WRITE, &
     H5FORTRAN_READ_ONLY_SERIAL => H5FORTRAN_READ_ONLY
 #endif
@@ -38,6 +39,7 @@ module h5fort
 #ifdef H5FORTRAN_SERIAL
   public :: h5fort_swrite, h5fort_sread, h5fort_sread_fixed, h5fort_swrite_attr
   public :: h5fort_write_attribute, h5fort_read_attribute
+  public :: h5fort_get_attribute_info
   public :: t_hdf5_attr, t_h5fort_serial
 #endif